filter remove cart

+++ shopping-cart-3d-frontend/src/components/CartCard.js
@@ -1,43 +1,60 @@
 import React, { Component } from 'react'
 import { connect } from 'react-redux'
-import { Typography, Row, Col, Icon } from 'antd'
+import { Typography, Row, Col, Icon, Button } from 'antd'
 import { removecart } from '../redux/actions/actions'
+let { Text } = Typography
 
 class CartCard extends Component {
 
-  handleDelete = () => {
-    this.props.removecart()
+  handleDelete = (id) => {
+    this.props.removecart(id)
     console.log(this.state)
+  }
+
+  switchRender(cartsList) {
+    if (cartsList.length == 0) {
+      return (
+        <Row style={{ width: '300px', padding: '7px 0' }}>
+          
+        </Row>
+      )
+    } else {
+      return (
+        <>
+          {
+            cartsList.map(cartItem => (
+              <li key={cartItem.id}>
+                <Row style={{ width: '300px', padding: '7px 0' }}>
+                  <Col span={9}>
+                    <img src={cartItem.image} style={{ width: '100px', height: '100px', border: '1px solid #000' }} />
+                  </Col>
+                  <Col span={12}>
+                    <Row>
+                      <Text>{cartItem.name}</Text>
+                    </Row>
+                    <Row>
+                      <Text>{cartItem.price}</Text>
+                    </Row>
+                  </Col>
+                  <Col span={3}>
+                    <Button type='link' onClick={() => this.handleDelete(cartItem.id)} key={cartItem.id}>
+                      <Icon type="delete" theme='filled' style={{ fontSize: '25px', color: '#ff4d4f' }} />
+                    </Button>
+                  </Col>
+                </Row>
+              </li>
+            ))
+          }
+        </>
+      )
+    }
   }
 
   render() {
     let cartsList = this.props.cartsList
-    let handleDelete = this.props
-    let { Text } = Typography
     return (
       <div>
-        {
-          cartsList.map(cartItem => (
-            <Row style={{ width: '300px', padding: '7px 0' }}>
-              <Col span={9}>
-                <img src={cartItem.image} style={{ width: '100px', height: '100px', border: '1px solid #000' }} />
-              </Col>
-              <Col span={14}>
-                <Row>
-                  <Text>{cartItem.name}</Text>
-                </Row>
-                <Row>
-                  <Text>{cartItem.price}</Text>
-                </Row>
-              </Col>
-              <Col span={1}>
-                <span style={{cursor: 'pointer'}} onClick={() => this.handleDelete()}>
-                  <Icon type="delete" theme ='filled'style={{fontSize:'25px',color:'#ff4d4f'}} />
-                </span>
-              </Col>
-            </Row>
-          ))
-        }
+        {this.switchRender(cartsList)}
       </div>
     )
   }
